Refactor and clean-up the code, also fix a bug.
When Postgres is running and we change its HBA file, we need to signal
Postgres that there are new rules to load (with a reload).

Also, share the otherNodesArray cache with the other procedures that are
needing it, namely the replication slot maintenance procedures. Finally,
refactor the node_active code to make it easier to maintain in the long run.